2012/05/07 fix ie7 ie9 firefox editor

+++ models/__init__.py
@@ -73,7 +73,7 @@ post_save.connect(create_user_profile, sender=User)
 User.add_to_class('views', models.PositiveIntegerField(default=0))
 User.add_to_class('responses', models.PositiveIntegerField(default=0))
 User.add_to_class('NewestQ', models.ForeignKey(Thread, related_name=_('the_newest_post'), blank=True, null=True, on_delete=models.SET_NULL))
-User.add_to_class('balance', models.PositiveIntegerField(default=0))
+User.add_to_class('balance', models.DecimalField(default=0,max_digits=8, decimal_places=2))
 
 User.add_to_class('gravatar', models.CharField(max_length=32))
 #User.add_to_class('has_custom_avatar', models.BooleanField(default=False))
